+++ spec/rails_app/app/controllers/sessions_controller.rb
@@ -1,0 +1,11 @@
+class SessionsController < ApplicationController
+  def signin
+    session[:admin] = true
+    redirect_to root_url, :notice => "Successfully signed in as admin"
+  end
+
+  def signout
+    session[:admin] = false
+    redirect_to root_url, :notice => "Successfully signed out"
+  end
+end