+++ spec/rails_app/app/models/post.rb
@@ -1,2 +1,4 @@
 class Post < ActiveRecord::Base
+  scope :published, lambda {where('published_at <= ?', Time.now)}
+  scope :unpublished, lambda {where('published_at > ?', Time.now)}
 end